agent_experiment/requirements.txt
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
torch>=1.9.0
networkx>=2.6.3
scipy>=1.7.0
statsmodels>=0.13.0
openai>=1.0.0
python-dotenv>=0.19.0
tqdm>=4.62.0
wandb==0.19.6